backend/django_project/requirements.txt
# Django and REST framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Environment and configuration
python-dotenv==1.0.0
django-environ==0.11.2

# Development
black==23.11.0
isort==5.12.0
mypy==1.7.0
django-stubs==4.2.6
djangorestframework-stubs==3.14.4

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0

# Type checking
types-python-dateutil==2.8.19.14
types-requests==2.31.0.10

# Utilities
requests==2.31.0
python-dateutil==2.8.2 
